First draft of the shell. Basic readline functionality and tokenizing.
Tokenizing reduces input to an array of token structs. A token struct
is a number representing a type, and then a union of data which depends
on that type. The idea is to emulate an algebraic datatype, like
data Token = Pipe | ChInp | ChOutp | Str String | ChOutApp | Background
           | GenOutRed FileDes | GenInOutRed FileDes FileDes | Rerun Int

Current functionality is supposed to just take the input string, tokenize,
and print out all the tokens found. For most tokens it works, but as of
now there is a bug that causes it to hang when passed a naked string.
Also on the todo list is to fix up strings with quotes around them to
*not* include the quotes in the array it is copied to.