Created additional parsing from tokens -> commands
The function in separate_commands in command.c takes a list of tokens
terminated by one with type EMPTY, and returns a pointer to an
array of commands. A command has a list of strings (to be given to
execlp), and 3 filedescriptors for input, output, and errors. In
addition to seperating along pipes, separate_commands translates
filenames (like in < "inp.txt" and ">> out.txt") to filedescriptors.
This is bordeline "doing too many things" for one function, but it
lends to a simplified definition of the command struct, as well
as making sure that all the child processes agree on file
descriptors (because the file descriptor table is cloned on forking).