We are close, but not quite (RIP > CS.limit)
Right now the code is in a clean state, mostly
what we want it to do. There is still a couple of printing issues
that we could take care of if we have enough space for them, but
those are not critical. Given that there are 9 minutes left to
the deadline, we have decided to take a late token for it.

The current problem right now is a mismatch between conceptual
model and in practice. We suspect that copying the sectors
in copy_loop somehow corrupts the memory that is supposed to
be the kernel. Copy loop behaves correctly, if the math is
correct. The portion that refers to the target is incremented
by 0x8000 each time (in agreement with the 0x40 sectors * 0x200 bytes read)
and the starting sector is incremented by 0x40 each time.
What has not been investigated is behavior after a large number of
iterations of the copy_loop, dwang suspects something funny might
be going on with the counter in al.
`pintos run alarm-zero` will execute, and the printouts reflect
it is actually getting through all the bootloader code but
I think it actually jumps *back* into it after when it should.
The actual error is "prefetch: [RIP > CS.limit]" which suggests
that we need to be more careful about segmentation