Created fixed point arithmetic
It should be fairly straightforward to use, and should act
like function invocations.
FRAC_BITS is the number of bits to the left of the decimal point
INT_BITS is the number to the right.
They add to 31 because there is also a sign bit.

Arithmetic options are available for combining two fixed point numbers
or a fixed point and an integer, based on choosing fp($OP) or fp($OP)int.