Implemented alarm clock
All tests pass. Could be made more efficient if we kept the list of
sleeping threads sorted, but we can do that later if necessary.
Used sleep_list (in thread.c) to keep track of sleeping threads, and
saved each thread's wakeup time inside of the thread. On each timer
interrupt, iterates through the sleeping threads and wakes up the ones
that are ready.