Added some page management to frame.c
It should now correctly create a `struct frame` when there is an
empty frame based on the results of palloc, otherwise it will
evict a frame and malloc no new frame struct.

It also should free that frame during writeback, but the
struct frame holding it should not be because it is probably
about to be used again. This will need more careful determination
of how to find an empty frame struct. Maybe the frame struct
should be freed.