Started file mappings
Currently, file mappings are implemented the same as file descriptors were,
with a dynamic array of file mappings. The uniqueness of mapping ids is
guaranteed in the same way as file descriptors. The only change is that
the dynamic array now contains a map_entry struct, defined in thread.h.
This struct contains an address that is the start of the mapping, as well
as the size of the file.
This deviates from the original hash table idea, but the efficiency is
comparable, and it's easier to implement (ie, through copy-paste).