Wrote the beginnings of unmap.
Need to recreate a frame_writeback function.
Loops through the pages for the file, and attempts to
writeback all that are in memory. Probably should
find someway to fully remove those entries from the page
table, or maybe just set invalid in the SPT.

Then deletes SPT entries

then closes the hidden file that keeps it open even if using process