Pretty sure I have the API I want for the filesystem cache
Basically the way it works is by wrapping the calls that inode.c
was doing to block_read and block_write. If it controls all the
access to the file system block device, then it can manage when
writes to the disk actually happen and be confident that it knows
what's up. Now I just need to figure out internals.
Passes the filesystem tests, except for occasionally syn-read and
syn-write