Caching actually works on the writing end.
No explicit writebacks after dirtying a block, just setting a dirty
flag. The cache also has the size of just one sector.
Note that the functions to set the dirty flags and so on
should be used, in case you have somebody that doesn't grok bitwise
arithmetic (*cough, cough* me).

Note that it passes filesys/base except for some of the syn ones,
I think. It seemed as though several had failed at one point but I
was unable to
reproduce those errors