Trying to not open removed files and the like
Changed dir from a block_sector_t to a struct inode * in
thread struct. This is a little bit more reversible, because
"inode_get_inumber" will always give the same answer for the
inode, but to go in the reverse direction requires an inode_open,
which scrubs the removed field (not desirable).

Still has a problem with creating a file in a directory that has
been deleted.